+++ NEWS.rst
@@ -6,8 +6,12 @@ every change, see the Git log.
 
 Latest
 ------
+* Patch: Fixed bug where assignment of a json object to a contained object would
+  cause a memory leak.
 * Patch: Fixed bug where re-assignment via the ``operator[]`` would cause a memory
   leak.
+* Minor: Added ``json::contains`` function for checking if a json object is contained
+  in another json object.
 
 10.0.1
 ------